Revert "diplom"
