--- README.md
@@ -1,30 +1,0 @@
-# Monkkee
-Тестирование сайта онлайн-дневника 
-https://my.monkkee.com
-
-# Чек лист:
-1. Тестирование логина
-2. Добавление заметки
-3. удаления заметки 
-4. редактирования заметки
-5. Проверить возможность создания нового аккаунта с верными данными пользователя.
-6. Проверить, что нельзя создать новый аккаунт без указания обязательных данных.
-7. Проверить возможность входа в аккаунт с верными данными пользователя.
-8. Проверить, что нельзя войти в аккаунт с неверными данными пользователя.
-9. Проверить возможность создания новой записи в дневнике с верными данными.
-10. Проверить, что нельзя создать новую запись в дневнике без указания названия.
-11. Проверить возможность редактирования существующей записи в дневнике с верными данными.
-12. Проверить, что нельзя редактировать запись в дневнике без правильных данных.
-13. Проверить возможность удаления существующей записи в дневнике с верными данными.
-14. Проверить, что нельзя удалить запись в дневнике без правильных данных.
-
-# **стек технологий**
-Maven
-TestNG
-JUnit
-Selenium
-Allure Reporting
-
-инструкции по  запуску тестов и генерации отчета
-login ivanchikaleksandr11@gmail.com
-pass alex1111
