--- src/main/java/Allure/AllureUtils.java
@@ -1,16 +1,0 @@
-package Allure;
-
-import org.openqa.selenium.OutputType;
-import org.openqa.selenium.TakesScreenshot;
-import org.openqa.selenium.WebDriver;
-
-public class AllureUtils {
-
-
-
-    //@Attachment(value = "Screenshot", type = "image/png")
-        public static byte[] takeScreenshot(WebDriver driver) {
-            return ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
-        }
-    }
-
