--- src/main/java/Page/BasePage.java
@@ -1,20 +1,0 @@
-package Page;
-
-import lombok.extern.log4j.Log4j2;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.time.Duration;
-
-@Log4j2
-public class BasePage {
-
-    protected WebDriver driver;
-    protected WebDriverWait wait;
-
-    public BasePage(WebDriver driver) {
-        this.driver = driver;
-        wait = new WebDriverWait(driver, Duration.ofSeconds(15));
-      log.info("Webdriver is {}",driver);
-    }
-}
