--- src/main/java/Page/DiaryEntry.java
@@ -1,38 +1,0 @@
-package Page;
-
-public class DiaryEntry  {
-    public DiaryEntry() {
-
-    }
-
-    public DiaryEntry(boolean b) {
-
-    }
-
-    public void setTitle() {
-
-    }
-
-    public void setContent() {
-    }
-
-    public String getTitle() {
-        return null;
-    }
-
-    public String getContent() {
-
-        return null;
-    }
-
-    public void delete() {
-
-    }
-
-    public boolean isErrorDisplayed() {
-        return false;
-    }
-
-
-
-}
