--- src/test/java/EditEntryWithoutPermissionTest.java
@@ -1,19 +1,0 @@
-import Page.DiaryEntry;
-import org.testng.annotations.Test;
-
-import static org.testng.AssertJUnit.assertTrue;
-
-public class EditEntryWithoutPermissionTest extends BaseTest {
-    @Test
-    public void EditEntryWithoutPermission() {
-        // Создание экземпляра класса, содержащего запись в дневнике без прав для редактирования.
-        DiaryEntry entry = new DiaryEntry(false);
-
-        // Редактирование заголовка и содержания записи.
-        entry.setTitle();
-        entry.setContent();
-
-        // Проверка отображения ошибки.
-        assertTrue(entry.isErrorDisplayed());
-    }
-}
