--- src/test/java/EditEntyTest.java
@@ -1,20 +1,0 @@
-import Page.DiaryEntry;
-import org.testng.annotations.Test;
-
-import static org.testng.AssertJUnit.assertEquals;
-
-public class EditEntyTest extends BaseTest {
-    @Test
-    public void EditEntry() {
-        // Создание экземпляра класса, содержащего запись в дневнике.
-        DiaryEntry entry = new DiaryEntry();
-
-        // Редактирование заголовка и содержания записи.
-        entry.setTitle();
-        entry.setContent();
-
-        // Проверка успешности редактирования записи.
-        assertEquals("Новый заголовок", entry.getTitle());
-        assertEquals("Новое содержание", entry.getContent());
-    }
-}
