+++ src/test/java/StartPage.java
@@ -1,0 +1,3 @@
+public class StartPage {
+
+}
